[refactor] 도메인의 의도를 좀 더 명확히 드러내기 위해 클래스 이름 변경 - LineResults -> LadderResults
