+++ src/main/java/nextstep/ladder/domain/game/LadderResults.java
@@ -1,18 +1,18 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.domain.game;
 
 import java.util.Collections;
 import java.util.List;
 
-public class LineResults {
+public class LadderResults {
 
     private static final String INVALID_POSITION = "결과를 가져올 수 없습니다. 위치를 다시 확인해주세요.";
     private final List<String> lineResults;
 
-    public LineResults(List<String> results) {
+    public LadderResults(List<String> results) {
         this.lineResults = results;
     }
 
-    public String match(int position) {
+    public String findBy(int position) {
         assertPositionRange(position);
         return lineResults.get(position);
     }
